web/requirements.txt
aldryn-apphooks-config==0.4.0
aldryn-style==1.1.4
dj-database-url==0.5.0
Django==1.11.11
django-appdata==0.2.0
django-classy-tags==0.8.0
django-cms==3.5.1
django-filer==1.3.0
django-formtools==2.1
django-mptt==0.8.7
django-polymorphic==1.3
django-sekizai==0.10.0
django-treebeard==4.2.1
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-installer==1.0.0
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-text-ckeditor==3.5.3
djangocms-video==2.0.4
easy-thumbnails==2.5
html5lib==0.9999999
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
six==1.11.0
tzlocal==1.5.1
Unidecode==0.4.21
